+++ examples/metamask/test/playwright/synpress.ts
@@ -1,5 +1,5 @@
-import { testWithSynpress } from '@synthetixio/synpress'
-import { metaMaskFixtures } from '@synthetixio/synpress/playwright'
+import { testWithSynpress } from '@synpress-cypress/synpress'
+import { metaMaskFixtures } from '@synpress-cypress/synpress/playwright'
 import connectedSetup from '../wallet-setup/connected.setup'
 
 export default testWithSynpress(metaMaskFixtures(connectedSetup))