+++ examples/metamask/test/wallet-setup/basic.setup.ts
@@ -1,5 +1,5 @@
-import { defineWalletSetup } from '@synthetixio/synpress'
-import { MetaMask } from '@synthetixio/synpress/playwright'
+import { defineWalletSetup } from '@synpress-cypress/synpress'
+import { MetaMask } from '@synpress-cypress/synpress/playwright'
 
 const SEED_PHRASE = 'test test test test test test test test test test test junk'
 const PASSWORD = 'SynpressIsAwesomeNow!!!'