+++ release/src/cypress/support/index.ts
@@ -1,5 +1,5 @@
 export {
   mockEthereum,
   synpressCommands as synpressCommandsForEthereumWalletMock
-} from '@synthetixio/ethereum-wallet-mock/cypress/support'
-export { synpressCommands as synpressCommandsForMetaMask } from '@synthetixio/synpress-metamask/cypress/support'
+} from '@synpress-cypress/ethereum-wallet-mock/cypress/support'
+export { synpressCommands as synpressCommandsForMetaMask } from '@synpress-cypress/synpress-metamask/cypress/support'