+++ release/src/playwright/index.ts
@@ -1,3 +1,3 @@
-export * from '@synthetixio/ethereum-wallet-mock/playwright'
-export * from '@synthetixio/synpress-metamask/playwright'
-export * from '@synthetixio/synpress-phantom/playwright'
+export * from '@synpress-cypress/ethereum-wallet-mock/playwright'
+export * from '@synpress-cypress/synpress-metamask/playwright'
+export * from '@synpress-cypress/synpress-phantom/playwright'