+++ wallets/ethereum-wallet-mock/src/cypress/support/mockEthereum.ts
@@ -19,7 +19,7 @@
  * @example
  * ```typescript
  * // In your Cypress support file
- * import { mockEthereum } from '@synthetixio/synpress';
+ * import { mockEthereum } from '@synpress-cypress/synpress';
  *
  * mockEthereum();
  * ```