+++ wallets/ethereum-wallet-mock/src/playwright/synpress.ts
@@ -1,4 +1,4 @@
-import { testWithSynpress } from '@synthetixio/synpress-core'
+import { testWithSynpress } from '@synpress-cypress/synpress-core'
 import { ethereumWalletMockFixtures } from '.'
 
 export default testWithSynpress(ethereumWalletMockFixtures)