+++ wallets/metamask/test/playwright/wallet-setup/connected.setup.ts
@@ -1,4 +1,4 @@
-import { defineWalletSetup } from '@synthetixio/synpress-cache'
+import { defineWalletSetup } from '@synpress-cypress/synpress-cache'
 import { MetaMask, getExtensionId } from '../../../src/playwright'
 
 const SEED_PHRASE = 'test test test test test test test test test test test junk'