+++ wallets/phantom/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @synthetixio/synpress-phantom
+# @synpress-cypress/synpress-phantom
 
 ## 0.0.12
 
@@ -6,5 +6,5 @@
 
 - Release
 - Updated dependencies
-  - @synthetixio/synpress-cache@0.0.12
-  - @synthetixio/synpress-core@0.0.12
+  - @synpress-cypress/synpress-cache@0.0.12
+  - @synpress-cypress/synpress-core@0.0.12